Refactored Duke into newDuke to run and parse
Refactored Duke into newDuke to run and parse
Repackaged as newDuke
